fix: 修复拼写错误和信号引用问题
refactor: 将DecimalField替换为FloatField以提升性能

fix: 增强数据加载和审计逻辑的健壮性

refactor: 优化筹码指标计算和衍生指标处理逻辑

fix: 修正动态阈值校准和风险评分逻辑

refactor: 统一使用终极信号替换旧信号引用

fix: 修复进攻层评分逻辑和配置处理

refactor: 简化风险惩罚评分逻辑并升级信号

fix: 增强任务上下文初始化和数据处理逻辑

refactor: 优化筹码健康分计算和保存流程

+++ stock_models/fund_flow.py
@@ -778,17 +778,17 @@ class BaseAdvancedFundFlowMetrics(models.Model):
         # --- 4.2 斜率指标 ---
         # 4.2.1 每日核心指标的斜率
         for name, verbose in CORE_METRICS.items():
-            vars()[f'{name}_slope_{p}d'] = models.DecimalField(max_digits=20, decimal_places=8, verbose_name=f'{verbose}{p}日斜率', null=True, blank=True)
+            vars()[f'{name}_slope_{p}d'] = models.FloatField(verbose_name=f'{verbose}{p}日斜率', null=True, blank=True)
         
         # 4.2.2 累计指标的斜率 (仅对周期>1)
         if p > 1:
             for name, verbose in CORE_METRICS.items():
                 if 'ratio' not in name and 'divergence' not in name:
-                    vars()[f'{name}_sum_{p}d_slope_{p}d'] = models.DecimalField(max_digits=22, decimal_places=8, verbose_name=f'{verbose}{p}日累计之{p}日斜率', null=True, blank=True)
+                    vars()[f'{name}_sum_{p}d_slope_{p}d'] = models.FloatField(verbose_name=f'{verbose}{p}日累计之{p}日斜率', null=True, blank=True)
 
         # --- 4.3 加速度指标 ---
         for name, verbose in CORE_METRICS.items():
-            vars()[f'{name}_accel_{p}d'] = models.DecimalField(max_digits=20, decimal_places=8, verbose_name=f'{verbose}{p}日加速度', null=True, blank=True)
+            vars()[f'{name}_accel_{p}d'] = models.FloatField(verbose_name=f'{verbose}{p}日加速度', null=True, blank=True)
 
     class Meta:
         abstract = True
